+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.3.0] / 2024-12-16
+### Features
+- `RevitInstallationUtils.InstalledRevit` without `componentGuid`.
+### Updates
+- Update `InstalledRevit` without `componentGuid`.
+
 ## [1.2.0] / 2024-05-23
 ### Features
 - Feature `ProductInfoUtils` to get `ProductInfo` using `msi.dll`.
@@ -68,6 +74,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - `ricaun.Revit.Installation.Tests`
 
 [vNext]: ../../compare/1.0.0...HEAD
+[1.3.0]: ../../compare/1.2.0...1.3.0
 [1.2.0]: ../../compare/1.1.2...1.2.0
 [1.1.2]: ../../compare/1.1.1...1.1.2
 [1.1.1]: ../../compare/1.1.0...1.1.1